KAN-17 Refactor Add and Update Service

+++ Wasla_Backend/Services/Implementation/DoctorServiceService.cs
@@ -64,20 +64,29 @@
             
             service.price = updateServiceDto.price;
 
-            _serviceDayRepo.RemoveRange(service.ServiceDays);
-            await _serviceDayRepo.SaveChangesAsync();
-            _serviceDateRepo.RemoveRange(service.ServiceDates);
-            await _serviceDateRepo.SaveChangesAsync();
-            _timeSlotRepo.RemoveRange(service.TimeSlots);
-            await _timeSlotRepo.SaveChangesAsync();
 
-            service.ServiceDays = updateServiceDto.serviceDays
+            if(updateServiceDto.serviceDays != null)
+            {
+                _serviceDayRepo.RemoveRange(service.ServiceDays);
+                await _serviceDayRepo.SaveChangesAsync();
+
+                service.ServiceDays = updateServiceDto.serviceDays
                 .Select(d => new ServiceDay { dayOfWeek = d.dayOfWeek })
                 .ToList();
+            }
+            
+            if(updateServiceDto.serviceDates != null)
+            {
+                _serviceDateRepo.RemoveRange(service.ServiceDates);
+                await _serviceDateRepo.SaveChangesAsync();
 
-            service.ServiceDates = updateServiceDto.serviceDates
-                .Select(d => new ServiceDate { date = d.date })
-                .ToList();
+                service.ServiceDates = updateServiceDto.serviceDates
+                    .Select(d => new ServiceDate { date = d.date })
+                    .ToList();
+            }
+
+            _timeSlotRepo.RemoveRange(service.TimeSlots);
+            await _timeSlotRepo.SaveChangesAsync();
 
             service.TimeSlots = updateServiceDto.timeSlots
                 .Select(t => new TimeSlot { start = t.start, end = t.end })
@@ -104,16 +113,20 @@
                 descriptionArabic = service.description.Arabic,
                 descriptionEnglish = service.description.English,
                 price = service.price,
-                serviceDays = service.ServiceDays.Select(day => new ServiceDayResponse
+                serviceDays = service.ServiceDays?
+                .Select(day => new ServiceDayResponse
                 {
                     id = day.id,
                     dayOfWeek = day.dayOfWeek,
-                }).ToList(),
-                serviceDates = service.ServiceDates.Select(date => new ServiceDateResponse
+                })
+                .ToList() ?? new List<ServiceDayResponse>(),
+                serviceDates = service.ServiceDates?
+                .Select(date => new ServiceDateResponse
                 {
                     id = date.id,
                     date = date.date,
-                }).ToList(),
+                })
+                .ToList() ?? new List<ServiceDateResponse>(),
                 timeSlots = service.TimeSlots.Select(slot => new TimeSoltsResponse
                 {
                     id = slot.id,
@@ -121,6 +134,7 @@
                     end = slot.end,
                 }).ToList()
             });
+        
         }
         public async Task DeleteServiceAsync(int serviceId)
         {
